dbの更新を1つのトランザクションにまとめた
